Add graphql support and show some product queries

+++ src/main/java/com/system/inventory/model/Product.java
@@ -19,7 +19,7 @@ public class Product {
 	private String id;
 	private String name;
 	private double price;
-	private double quantity;
+	private int quantity;
 	
 	@ManyToOne
 	private Category category;
